# Final iteration of performance tweaks