--- a/lib/index.js
+++ b/lib/index.js
@@ -8,8 +8,10 @@ const placeholders = require('./placeholders')
 let delimitersSettings = []
 let conditionals, loops, scopes, data
 
+const matchOperatorsRegexp = /[|\\{}()[\]^$+*?.]/g
+
 function escapeRegexpString (input) {
-  return input.replace(/[|\\{}()[\]^$+*?.]/g, '\\$&')
+  return input.replace(matchOperatorsRegexp, '\\$&')
 }
 
 module.exports = function PostHTMLExpressions (options) {