--- a/lib/index.js
+++ b/lib/index.js
@@ -176,16 +176,26 @@ function walk (opts, nodes) {
         throw new Error('You must provide at least one loop argument')
       }
 
+      // converts nodes to a string. These nodes will be changed within the loop
+      const treeString = JSON.stringify(node.content)
+      const keys = loopParams.keys
+
+      // creates a copy of the keys that will be changed within the loop
+      const localsBackup = makeBackupLoopKeys(keys[0], keys[1], opts.locals)
+
       // run the loop, different types of loops for arrays and objects
       if (Array.isArray(target)) {
         for (let index = 0; index < target.length; index++) {
-          m.push(executeLoop(loopParams.keys, target[index], index, node))
+          m.push(executeLoop(keys, target[index], index, opts.locals, treeString))
         }
       } else {
         for (let key in target) {
-          m.push(executeLoop(loopParams.keys, target[key], key, node))
+          m.push(executeLoop(keys, target[key], key, opts.locals, treeString))
         }
       }
+
+      // returns the original keys values that was changed within the loop
+      opts.locals = revertBackupedLocals(keys[0], keys[1], opts.locals, localsBackup)
 
       // return directly out of the loop, which will skip the "each" tag
       return m
@@ -259,18 +269,44 @@ function parseLoopStatement (input) {
 }
 
 /**
+ * Creates a backup of keys within the loop
+ */
+function makeBackupLoopKeys (p1, p2, locals) {
+  let backup = {}
+  if (p1 && data.locals.hasOwnProperty(p1)) backup[p1] = data.locals[p1]
+  if (p2 && data.locals.hasOwnProperty(p2)) backup[p2] = data.locals[p2]
+
+  return backup
+}
+
+/**
+ * Returns the original keys values
+ */
+function revertBackupedLocals (p1, p2, locals, backup) {
+  // Remove loop keys from locals
+  delete locals[p1]
+  delete locals[p2]
+
+  // Revert copied keys
+  if (p1 && backup.hasOwnProperty(p1)) locals[p1] = backup[p1]
+  if (p2 && backup.hasOwnProperty(p2)) locals[p2] = backup[p2]
+
+  return locals
+}
+
+/**
  * Creates a set of local variables within the loop, and evaluates all nodes
  * within the loop, returning their contents
  */
-function executeLoop (loopParams, p1, p2, node) {
+function executeLoop (loopParams, p1, p2, locals, treeString) {
   // two loop locals are allowed
   // - for arrays it's the current value and the index
   // - for objects, it's the value and the key
-  const scopedLocals = {}
+  const scopedLocals = locals
   scopedLocals[loopParams[0]] = p1
   if (loopParams[1]) scopedLocals[loopParams[1]] = p2
 
-  return executeScoped(scopedLocals, node)
+  return walk({ locals: scopedLocals }, JSON.parse(treeString))
 }
 
 /**