--- a/lib/index.js
+++ b/lib/index.js
@@ -1,7 +1,6 @@
 'use strict'
 
 const vm = require('vm')
-const merge = require('lodash.merge')
 const cloneDeep = require('lodash.clonedeep')
 const placeholders = require('./placeholders')
 
@@ -215,7 +214,15 @@ function walk (opts, nodes) {
         throw new Error('You must provide an object to make scope')
       }
 
-      m.push(executeScoped(target, node))
+      const keys = Object.keys(target)
+
+      // creates a copy of the keys that will be changed within the loop
+      const localsBackup = makeLocalsBackup(keys, opts.locals)
+
+      m.push(executeScoped(target, opts.locals, node))
+
+      // returns the original keys values that was changed within the loop
+      opts.locals = revertBackupedLocals(keys, opts.locals, localsBackup)
 
       // return directly out of the loop, which will skip the "scope" tag
       return m
@@ -316,11 +323,7 @@ function executeLoop (loopParams, p1, p2, locals, treeString) {
 /**
  * Runs walk function with arbitrary set of local variables
  */
-function executeScoped (scopedLocals, node) {
-  // merge nondestructively into existing locals
-  const scopedOptions = merge(cloneDeep(data), { locals: scopedLocals })
-  // walk through the contents and run replacements with modified options
-  // we need to clone the node because the normal operation modifies
-  // the node directly
-  return walk(scopedOptions, cloneDeep(node.content))
+function executeScoped (scopeLocals, locals, node) {
+  const scopedLocals = Object.assign(locals, scopeLocals)
+  return walk({ locals: scopedLocals }, node.content)
 }