--- a/lib/index.js
+++ b/lib/index.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const vm = require('vm')
-const cloneDeep = require('lodash.clonedeep')
+const cloneDeep = require('fclone')
 const placeholders = require('./placeholders')
 
 let delimitersSettings = []