--- a/lib/index.js
+++ b/lib/index.js
@@ -5,7 +5,7 @@ const cloneDeep = require('fclone')
 const placeholders = require('./placeholders')
 
 let delimitersSettings = []
-let conditionals, loops, scopes, data
+let conditionals, loops, scopes
 
 const matchOperatorsRegexp = /[|\\{}()[\]^$+*?.]/g
 
@@ -23,9 +23,6 @@ module.exports = function PostHTMLExpressions (options) {
     scopeTags: ['scope'],
     locals: {}
   }, options)
-
-  // make data for expressions
-  data = { locals: options.locals }
 
   // set tags
   loops = options.loopTags
@@ -59,7 +56,7 @@ module.exports = function PostHTMLExpressions (options) {
   }
 
   // kick off the parsing
-  return walk.bind(null, data)
+  return walk.bind(null, { locals: options.locals })
 }
 
 function walk (opts, nodes) {
@@ -283,7 +280,7 @@ function makeLocalsBackup (keys, locals) {
 
   for (let i = 0; i < keys.length; i++) {
     const key = keys[i]
-    if (data.locals.hasOwnProperty(key)) backup[key] = cloneDeep(data.locals[key])
+    if (locals.hasOwnProperty(key)) backup[key] = cloneDeep(locals[key])
   }
 
   return backup