--- a/lib/placeholders.js
+++ b/lib/placeholders.js
@@ -2,14 +2,17 @@
 
 const vm = require('vm')
 
+const matchHtmlRegexp = /[&<>"']/g
+const htmlSymbolsMap = {
+  '&': '&amp;',
+  '<': '&lt;',
+  '>': '&gt;',
+  '"': '&quot;',
+  '\'': '&#039;'
+}
+
 function escapeHtmlCharacters (unsafe) {
-  return unsafe.replace(/[&<>"']/g, (match) => ({
-    '&': '&amp;',
-    '<': '&lt;',
-    '>': '&gt;',
-    '"': '&quot;',
-    '\'': '&#039;'
-  })[match])
+  return unsafe.replace(matchHtmlRegexp, (match) => htmlSymbolsMap[match])
 };
 
 module.exports = function replace (input, ctx, settings) {