--- a/package.json
+++ b/package.json
@@ -10,8 +10,7 @@
     "url": "https://github.com/posthtml/posthtml-expressions/issues"
   },
   "dependencies": {
-    "lodash.clonedeep": "^4.3.2",
-    "lodash.merge": "^4.4.0"
+    "lodash.clonedeep": "^4.3.2"
   },
   "devDependencies": {
     "ava": "^0.16.0",