--- a/test/fixtures/scope.expected.html
+++ b/test/fixtures/scope.expected.html
@@ -1,5 +1,6 @@
-<p>x</p>
+<p>x Scope x</p>
 
   <p>Hi. I am John. I am 26 years old.</p>
+  <p>test</p>
 
-<p>x</p>
+<p>x Scope x</p>