--- a/test/fixtures/scope.html
+++ b/test/fixtures/scope.html
@@ -1,5 +1,6 @@
-<p>x</p>
-<scope with="data">
+<p>x {{name}} x</p>
+<scope with="author">
   <p>Hi. I am {{name}}. I am {{age}} years old.</p>
+  <p>{{key}}</p>
 </scope>
-<p>x</p>
+<p>x {{name}} x</p>