--- a/test/fixtures/scope_nested.expected.html
+++ b/test/fixtures/scope_nested.expected.html
@@ -1,0 +1,11 @@
+<p>global</p>
+
+  <p>scope</p>
+
+    <p>one</p>
+
+
+    <p>two</p>
+
+
+<p>global</p>