--- a/test/fixtures/scope_nested.html
+++ b/test/fixtures/scope_nested.html
@@ -1,0 +1,11 @@
+<p>{{key}}</p>
+<scope with="scope">
+  <p>{{key}}</p>
+  <scope with="one">
+    <p>{{key}}</p>
+  </scope>
+  <scope with="two">
+    <p>{{key}}</p>
+  </scope>
+</scope>
+<p>{{key}}</p>