## Proposed Changes

Fix for #26.

### Benchmark [only 5x loop]

| before | after    | faster |
|--------|----------|--------|
| 8 op/s | 111 op/s | 13.8x  |

### Benchmark [complex from [posthtml-benchmark](https://github.com/mrmlnc/posthtml-benchmark)]

| before | after   | faster |
|--------|---------|--------|
| 7 op/s | 61 op/s | 8.7x   |

## Types of Changes

Fully compatible with the current version.