@michael-ciniawsky 2016-11-18
@mrmlnc I update at least the README (already in progress (90%)) and make new release. Afterwards I add you as npm owner aswell. (we need a npm release bot for posthtml org 😛, but off-topic)