@mrmlnc 2016-11-18
@michael-ciniawsky,

I think is better to not give me the rights for plugins in npm, because I prefer to create PR 😺