@michael-ciniawsky 2016-11-18
Doesn't hurt for asap bug fixes 😛 , e.g in case i'm busy with social work etc., making PR for features is good practice 👍 nevertheless + do as you prefer 😎